pyproject.toml
[tool.poetry]
name = "sarafan2"
version = "0.1.0"
description = ""
authors = ["ika <mamina747@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
